computed propertie

+++ app/Livewire/ArticleList.php
@@ -5,6 +5,7 @@ namespace App\Livewire;
 use App\Models\Article;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -19,6 +20,24 @@ class ArticleList extends AdminComponent
     public $showOnlyPublished = false;
 
     public string $search = '';
+
+    #[Computed()]
+    public function articles()
+    {
+        $query = Article::query();
+
+        if ($this->showOnlyPublished) {
+            $query->where('published', 1);
+        }
+
+        if (trim($this->search) !== '') {
+            $s = '%' . str_replace('%', '\\%', $this->search) . '%';
+            $query->where('title', 'like', $s);
+        }
+
+        return $query->paginate(10, pageName: 'article-page');
+
+    }
 
 
     public function showAll(): void
@@ -39,29 +58,34 @@ class ArticleList extends AdminComponent
         $this->resetPage(pageName: 'article-page');
     }
 
-    public function render(): Factory|View|\Illuminate\View\View
-    {
-        $query = Article::query();
+//    public function render(): Factory|View|\Illuminate\View\View
+//    {
+//        return view('livewire.article-list');
+////        $query = Article::query();
+////
+////        if ($this->showOnlyPublished) {
+////            $query->where('published', 1);
+////        }
+////
+////        if (trim($this->search) !== '') {
+////            $s = '%' . str_replace('%', '\\%', $this->search) . '%';
+////            $query->where('title', 'like', $s);
+////        }
+////
+////        return view('livewire.article-list', [
+////            'articles' => $query->paginate(10, pageName: 'article-page')
+////        ]);
+//    }
 
-        if ($this->showOnlyPublished) {
-            $query->where('published', 1);
-        }
-
-        if (trim($this->search) !== '') {
-            $s = '%' . str_replace('%', '\\%', $this->search) . '%';
-            $query->where('title', 'like', $s);
-        }
-
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10, pageName: 'article-page')
-        ]);
-    }
-
-    public function delete(int $id): void
+    public function delete(Article $article): void
     {
         try {
-            $article = Article::findOrFail($id);
+            if ($this->articles()->count() < 10){
+                throw new \Exception('Nope');
+            }
             $article->delete();
+            unset($this->articles);
+            cache()->forget('count-published-articles');
             $this->dispatch('notify', type: 'success', message: 'Article deleted successfully.');
         } catch (\Throwable $e) {
             $this->dispatch('notify', type: 'error', message: 'Failed to delete article.');
